added functionality for comparing outputs and a batch file for making a complete MLP run

+++ Models/C_Models/fnlib.h
@@ -3,8 +3,8 @@
 
 
 typedef int reg;
-typedef float* mat;
-typedef float* vec;
+typedef double* mat;
+typedef double* vec;
 typedef int acttype;
 
 void load_m(reg a, void * b, int rows, int cols);
